Add some functions for fetching scan results.
get_plugin_output() takes a scan name and a plugin ID and fills in a
dict with the output of that plugin ID.

get_host_vulns() is a helper function used by get_plugin_output(). It
fills in a dict containing host vulnerabilites, keyed by scan id and
host id.

scan_details() is a helper function used by get_host_vulns(). It takes
the name of a scan and fetches the details for it.